break out in-mem store to separate file
